+++ src/Synapse/Synapse/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 
 ## Upcoming Release
+
+## Version 3.0.9
 * Updated Azure.Analytics.Synapse.Artifacts to 1.0.0-preview.20.
 * Fixed the issue that Azure.Core.AccessToken is used before assigned.